feat: lis3dh: Add temperature channel
Read out the temperature sensor through the SENSOR_CHAN_TEMP channel
of the sensor API.
The measured temperature is relative to a uncalibrated one.

According to the datasheet, the ADC has a resolution of 10 bit in
normal mode and 8 bit in low-power mode, while the temperature sensor
is specified with a 8-bit resolution. The LSb of the left-aligned
measurement is in both power modes identical (bit 6).

At this point it is unclear where exactly the sign bit of the readout
lies. It is therefore very likely that it doesn't work correctly for
negative values.