feat: lis3dh: Disable interrupt functionality on invalid callback
Being able to disable the interrupt functionality is useful for sleep
states, there one might not want to be waken from. Rather deal with
the interrupt if awake.

So in order to disable the interrupt, the callback can be set to NULL
in userspace. This reconfigures the pin to a simple input pin.
To enable interrupts, set the callback to a valid function pointer.