feat: tests: enable akita_gen2 for nffs test build
This patch changes sources and buildsystem files, so that akita_gen2 can
be used as target board. We have to change guards in some test
implementations, add power setting to the main file, add empty LPM
kernel hooks, add logging settings to the prj.conf and fix the
CMakeLists file to properly set parameters for the NFFS flash area.

Signed-off-by: Franco Saworski <franco.saworski@blik.io>