feat: soc: add flash readout protection
This patch changes the flash configuration so that it is not possible to
dump the flash memory contents and read out the binary.

The JLink Base tool errors out when trying to read the memory.

However, connecting to the device with GDB, you can still dump the
binary. The flash configuration flag in the binary is then unset. It is
unclear why.

Signed-off-by: Franco Saworski <franco.saworski@blik.io>