fix: kw41z: remove K_FOREVER in kw41z_cca sem_take
This patch puts k_sem_take(&sem, K_FOREVER) in a while loop and
changes the timeout parameter to a constant time. This prevents
the kernel from descheduling the thread and going to low power
during CCA in kw41z_cca().

Before this patch, Akita woke up twice per ten second TX interval,
which allowed a race condition, because during CCA's k_sem_take the
system would go to lowpower and wake up after CCA is done. On CCA's
deepsleep wakeup, the radio ISR interrupted _sys_soc_suspend before
the system can change the DCDC configuration. This leads to an
instable system, with no radio or networking code executing anymore.
The exact details are unknown.

The problem can be reproduced with inserting a >3ms delay before
lpm_1V5_1V8_stepping_activate(). One occurence of this delay or
race condition is enough to destabilize the system forever until
hard system reset.

The watchdog did not catch this problem, because it is reset in the
RTC ISR.

This problem was found during investigation of a potential race
condition in _sys_soc_suspend and the field test problems.

For some reason, the CCA interrupt is able to wakeup the system from
LLS3, even though our power.c implementation only enables RTC in LLWU.
The CCA wakeup made the system operable to begin with, though unstable.

A similar problem was found in kw41z_tx() and fixed in blik-v1.0.4,
commit-id 7a4afd28 in September. TX was not able to wakeup the system,
which led to delayed transmissions after the RTC woke up the system
again ten seconds later.

Signed-off-by: Franco Saworski <franco.saworski@blik.io>